GDB-12170: Separate authentication from authorization
## What
Refactor the authentication logic to introduce a new AuthorizationService, separating concerns between authentication and authorization.

## Why
This change enhances code maintainability and clarity by clearly delineating authentication responsibilities from authorization, allowing for easier future modifications and improvements.

## How
- Created a new `AuthorizationService` to handle authorization-related operations.
- Updated existing code to replace instances of `AuthenticationService` with `AuthorizationService` where appropriate.
- Refactored methods to utilize the new service for checking permissions and roles.
- Removed redundant authentication checks from various components and services.
