+++ packages/api/src/services/security/index.ts
@@ -8,3 +8,4 @@ export {SecurityContextService} from './security-context.service';
 export {OpenidConfigMapper} from './openid-config.mapper';
 export {AuthoritiesUtil} from './utils/authorities-util';
 export {GrantedAuthoritiesUiModelMapper} from './mappers/granted-authorities-ui-model.mapper';
+export {AuthorizationService} from './authorization.service';
